+++ firmware/main/api/routes/system.h
@@ -6,5 +6,6 @@ extern const httpd_uri_t route_get_system;
 extern const httpd_uri_t route_post_reboot;
 extern const httpd_uri_t route_get_system_hardware;
 extern const httpd_uri_t route_put_system_hardware;
+extern const httpd_uri_t route_get_system_gpio;
 extern const httpd_uri_t route_get_system_log;
 extern const httpd_uri_t route_post_system_log_clear;